refactor(api): split router into domain handlers
